4.23.26 - The "We Tested More Things" Update - BoardMeeting:   - Simplified the DriveBase code. - AutoCommander:   - IntakeArtifactCommand -> IntakeArtifactsCommand   - Made use of commandQueue.add() returning a boolean.   - ...Now checks if the command queue (which is now an ArrayList) has commands in it before doing anything with them. - TeamCode:   - Started work on autonomous pathing.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoCommander/BotCommand.java
@@ -6,8 +6,8 @@ import org.firstinspires.ftc.teamcode.boardmeeting.ProgramBoard;
 
 public class BotCommand {
     public ProgramBoard board;
-    public float time;
-    public BotCommand(ProgramBoard board, float sleepTime) {
+    public double time;
+    public BotCommand(ProgramBoard board, double sleepTime) {
         this.board = board;
         this.time = sleepTime;
     }
